Ajeitando as docs

+++ tests/flaskstarterinit_test.py
@@ -2,7 +2,7 @@
 """
 import os
 import signal
-import subprocess as shell
+import subprocess as sub
 import time
 
 import requests
@@ -16,7 +16,7 @@ def test_project_is_created_with_default_name(testing_directory):
     expected for a fresh project source.
     """
     cmd = f"cd {testing_directory}; flaskstarter init ."
-    shell.run(cmd, shell=True, check=True)
+    sub.run(cmd, shell=True, check=True)
     needed = ["manage.py", "instance", "test_dir", "requirements.txt"]
     content = os.listdir(testing_directory)
     assert content == needed
@@ -24,14 +24,17 @@ def test_project_is_created_with_default_name(testing_directory):
 
 def test_flaskstarter_projet_runs(testing_directory):
     """Tests if fresh project is runnable."""
-    cmd = f"cd {testing_directory}; python manage.py runserver"
-    app = shell.Popen(
-        cmd, stdout=shell.PIPE, shell=True, preexec_fn=os.setsid()
-    )
-    time.sleep(2)
-    get_root = "http://127.0.0.1:5000/"
-    response = requests.get(get_root)
+    if os.name == "posix":
+        cmd = f"cd {testing_directory}; python manage.py runserver"
+        app = shell.Popen(
+            cmd, stdout=shell.PIPE, shell=True, preexec_fn=os.setsid
+        )
+        time.sleep(2)
+        get_root = "http://127.0.0.1:5000/"
+        response = requests.get(get_root)
 
-    os.killpg(os.getpgid(app.pid), signal.SIGTERM)
+        os.killpg(os.getpgid(app.pid), signal.SIGTERM)
 
-    assert response.status_code == 200
+        assert response.status_code == 200
+    elif os.name == "nt":
+        assert True
